requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-multipart
requests
beautifulsoup4
apscheduler
pandas
openpyxl
matplotlib
sqlalchemy